+++ Packages/com.unity.inputsystem/InputSystem/Devices/InputDeviceBuilder.cs
@@ -392,9 +392,13 @@ namespace UnityEngine.InputSystem.Layouts
             // Set flags and misc things.
             control.noisy = controlItem.isNoisy;
             control.synthetic = controlItem.isSynthetic;
+            control.usesStateFromOtherControl = !string.IsNullOrEmpty(controlItem.useStateFrom);
+            control.dontReset = (control.noisy || controlItem.dontReset) && !control.usesStateFromOtherControl; // Imply dontReset for noisy controls.
             if (control.noisy)
                 m_Device.noisy = true;
             control.isButton = control is ButtonControl;
+            if (control.dontReset)
+                m_Device.hasDontResetControls = true;
 
             // Remember the display names from the layout. We later do a proper pass once we have
             // the full hierarchy to set final names.
@@ -414,8 +418,7 @@ namespace UnityEngine.InputSystem.Layouts
                 control.m_MaxValue = controlItem.maxValue;
 
             // Pass state block config on to control.
-            var usesStateFromOtherControl = !string.IsNullOrEmpty(controlItem.useStateFrom);
-            if (!usesStateFromOtherControl)
+            if (!control.usesStateFromOtherControl)
             {
                 control.m_StateBlock.byteOffset = controlItem.offset;
                 control.m_StateBlock.bitOffset = controlItem.bit;
@@ -583,6 +586,7 @@ namespace UnityEngine.InputSystem.Layouts
             // Copy its state settings.
             child.m_StateBlock = referencedControl.m_StateBlock;
             child.usesStateFromOtherControl = true;
+            child.dontReset = referencedControl.dontReset;
 
             // At this point, all byteOffsets are relative to parents so we need to
             // walk up the referenced control's parent chain and add offsets until
@@ -872,10 +876,10 @@ namespace UnityEngine.InputSystem.Layouts
                 throw new NotSupportedException($"Device '{m_Device}' exceeds maximum supported control count of {1U << InputDevice.kControlIndexBits} (has {m_Device.allControls.Count} controls)");
 
             // Device is not in m_ChildrenForEachControl so use index -1.
-            FinalizeControlHierarchyRecursive(m_Device, -1, m_Device.m_ChildrenForEachControl);
+            FinalizeControlHierarchyRecursive(m_Device, -1, m_Device.m_ChildrenForEachControl, false, false);
         }
 
-        private void FinalizeControlHierarchyRecursive(InputControl control, int controlIndex, InputControl[] allControls)
+        private void FinalizeControlHierarchyRecursive(InputControl control, int controlIndex, InputControl[] allControls, bool noisy, bool dontReset)
         {
             // Make sure we're staying within limits on state offsets and sizes.
             if (control.m_ChildCount == 0)
@@ -898,6 +902,19 @@ namespace UnityEngine.InputSystem.Layouts
             SetDisplayName(control, displayNameFromLayout, shortDisplayNameFromLayout, false);
             SetDisplayName(control, displayNameFromLayout, shortDisplayNameFromLayout, true);
 
+            if (control != control.device)
+            {
+                if (noisy)
+                    control.noisy = true;
+                else
+                    noisy = control.noisy;
+
+                if (dontReset)
+                    control.dontReset = true;
+                else
+                    dontReset = control.dontReset;
+            }
+
             // Recurse into children. Also bake our state offset into our children.
             var ourOffset = control.m_StateBlock.byteOffset;
             var childCount = control.m_ChildCount;
@@ -908,7 +925,7 @@ namespace UnityEngine.InputSystem.Layouts
                 var child = allControls[childIndex];
                 child.m_StateBlock.byteOffset += ourOffset;
 
-                FinalizeControlHierarchyRecursive(child, childIndex, allControls);
+                FinalizeControlHierarchyRecursive(child, childIndex, allControls, noisy, dontReset);
             }
 
             control.isSetupFinished = true;